Add axios dependency, implement volunteering registration, and enhance response voting logic

+++ Backend/src/utils/cronJobs.ts
@@ -4,22 +4,26 @@ import trashPartner from "../models/trash/trashPartnerModel";
 import Coupon from "../models/couponModel";
 import trashCoupon from "../models/trash/trashCouponModel";
 import Token from "../models/tokenModel";
-
+import Response from "../models/responseModel";
+import Report from "../models/reportModel";
+import ResponseService from "../services/responseService";
 
 cron.schedule("0 0 * * *", async () => {
     try {
         console.log("Starting expired partners cleanup task...");
         const expiredPartners = await Partner.find({ endDate: { $lte: new Date() } });
 
-        await Promise.all(
-            expiredPartners.map(async (partner) => {
-                partner.hasExpired = true;
-                await trashPartner.create({ ...partner.toObject() });
-                await partner.deleteOne();
-            })
-        );
+        if (expiredPartners.length > 0) {
+            await Promise.all(
+                expiredPartners.map(async (partner) => {
+                    partner.hasExpired = true;
+                    await trashPartner.create({ ...partner.toObject() });
+                    await partner.deleteOne();
+                })
+            );
+        }
 
-        console.log("Expired partners cleanup task completed.");
+        console.log(`${expiredPartners.length} Expired partners cleanup task completed.`);
     } catch (error) {
         console.error("Error in expired partners cleanup task:", error);
     }
@@ -35,14 +39,16 @@ cron.schedule("5 0 * * *", async () => {
 
         const expiredCoupons = await Coupon.find({ expiryDate: { $lte: oneWeekFromNow } });
 
-        await Promise.all(
-            expiredCoupons.map(async (coupon) => {
-                await trashCoupon.create({ ...coupon.toObject() });
-                await coupon.deleteOne();
-            })
-        );
+        if (expiredCoupons.length > 0) {
+            await Promise.all(
+                expiredCoupons.map(async (coupon) => {
+                    await trashCoupon.create({ ...coupon.toObject() });
+                    await coupon.deleteOne();
+                })
+            );
+        }
 
-        console.log("Expired coupons cleanup task completed.");
+        console.log(`${expiredCoupons.length}Expired coupons cleanup task completed.`);
     } catch (error) {
         console.error("Error in expired coupons cleanup task:", error);
     }
@@ -59,3 +65,4 @@ cron.schedule("10 0 * * *", async () => {
         console.error("Error in expired tokens cleanup task:", error);
     }
 });
+
